controllers/github/secret_scanning: Move `GitHubPublicKey` structs to the `github` module
The `github` and `tests/util/github` modules should not need to import code from the HTTP layer or `controllers` module. This reduces the coupling and should allow us to extract the GitHub code into a dedicated subpackage.
